Modified article generation functional. Created base stratagy for article Generation
