+++ src/ArticleGeneration/Strategy/BaseStrategy.php
@@ -1,0 +1,101 @@
+<?php
+
+namespace App\ArticleGeneration\Strategy;
+
+use App\ArticleGeneration\ArticleGenerationInterface;
+use App\ArticleGeneration\PromotedWord\PromotedWordInserter;
+use App\Entity\Module;
+use App\Repository\ModuleRepository;
+use Faker\Factory;
+use Twig\Environment;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Error\SyntaxError;
+
+/**
+ * Базовая стратегия для генерации статьи
+ */
+abstract class BaseStrategy implements ArticleGenerationInterface
+{
+    /**
+     * Сервис вставки продвигаемых слов
+     */
+    private PromotedWordInserter $wordInserter;
+
+    /**
+     * Сервис для рендеринга контента в шаблон
+     */
+    private Environment $twig;
+
+    private ModuleRepository $moduleRepository;
+
+    public function __construct(
+        PromotedWordInserter $wordInserter,
+        Environment          $twig,
+        ModuleRepository     $moduleRepository
+    ) {
+        $this->wordInserter = $wordInserter;
+        $this->twig = $twig;
+        $this->moduleRepository = $moduleRepository;
+    }
+
+    /**
+     * Заполняет плейсхолдеры сгенерированным текстом
+     *
+     * @param Module[] $modules
+     * @return array - массив тел модулей с заполненными плейсхолдерами
+     * @throws LoaderError
+     * @throws RuntimeError
+     * @throws SyntaxError
+     */
+    protected function fillPlaceholders(array $modules): array
+    {
+        $faker = Factory::create();
+        $articleBody = [];
+        foreach ($modules as $module) {
+            $data = [];
+            if (preg_match('/{{(\s)*?title?(\|raw)?(\s)*?}}/', $module->getBody())) {
+                $data['title'] = $faker->sentence();
+            }
+
+            // Todo В принципе p почти везде одинаков, значит можно сделать вставку чисто в текст параграфов. Остальное мимо
+            if (preg_match('/{{(\s)*?paragraph?(\|raw)?(\s)*?}}/', $module->getBody())) {
+                $data['paragraph'] = $faker->paragraph(rand(1, 10));
+            }
+
+            if (preg_match('/{{(\s)*?paragraphs?(\|raw)?(\s)*?}}/', $module->getBody())) {
+                $data['paragraphs'] = '';
+                foreach ($faker->paragraphs(rand(2, 7)) as $paragraph) {
+                    $data['paragraphs'] .= PHP_EOL . '<p class="lead mb-0">' . $paragraph . '</p>';
+                }
+            }
+            // ToDo Добавить imagePath после того как простоим файловую систему
+            $articleBody[] = $this->getTwig()->render('article/components/article_module.html.twig', [
+                'data' => $data,
+                'module' => $module
+            ]);
+        }
+
+        return $articleBody;
+    }
+
+    /**
+     * Базовый абстрактный метод генерации статьи
+     */
+    public abstract function generate(object $articleDTO);
+
+    protected function getTwig(): Environment
+    {
+        return $this->twig;
+    }
+
+    protected function getWordInserter(): PromotedWordInserter
+    {
+        return $this->wordInserter;
+    }
+
+    protected function getModuleRepository(): ModuleRepository
+    {
+        return $this->moduleRepository;
+    }
+}
